+++ src/Plugin/Trajectories.cs
@@ -104,23 +104,11 @@ namespace Trajectories
             FlightOverlay.Update();
             NavBallOverlay.Update();
             MainGUI.Update();
-            if (!Settings.NewGui)
-#pragma warning disable CS0618 // Type or member is obsolete
-                OldGUI.Update();
-#pragma warning restore CS0618 // Type or member is obsolete
         }
 
-#if DEBUG && DEBUG_TELEMETRY
+#if DEBUG_TELEMETRY
         internal void FixedUpdate() => Trajectory.DebugTelemetry();
 #endif
-
-        internal void OnGUI()
-        {
-            if (!Settings.NewGui)
-#pragma warning disable CS0618 // Type or member is obsolete
-                OldGUI.OnGUI();
-#pragma warning restore CS0618 // Type or member is obsolete
-        }
 
         internal void OnDestroy()
         {